Create bitcointransaction_scraper.py
This scraper, scrapes unconfirmed transactions from a blockchain website. It lists all the variables and then looks to see what the largest transaction value is for that minute. The largest transaction is then sent to Redis' caching mechanism, which holds it for a minute. After this minute, it is sent to a collection in mongo database.